Completed following exercise from Copilot, adding all functionality.: Exercise: Extend the Vehicle Program Add a new subclass:
Create a new subclass of Vehicle called Truck.
The Truck class should have an engineType attribute and a cargoCapacity attribute.
The Truck class should have a drive method that prints a message including the engineType and speed.
Modify the Vehicle class:

Add a fuel attribute to the Vehicle class.
Modify the drive method in the Vehicle class to decrease the fuel by 1 unit each time the drive method is called.
If the fuel is 0 or less, print a message indicating that the vehicle is out of fuel and cannot drive.
Test the new functionality:

Create instances of the Truck class with different engineType and cargoCapacity.
Set the fuel attribute for each vehicle.
Call the drive method multiple times and observe the fuel decreasing.
Ensure that the vehicle stops driving when out of fuel.